feat(knowledgebase): refactor knowledge base routes and components, add document and chunk tabs
